Switched to xproj, removed serialization and intern
