+++ Semver.sln
@@ -1,15 +1,22 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Semver", "Semver\Semver.csproj", "{95B2542C-E371-4AB4-9E2D-B2665CBBD629}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Semver.Test", "Semver.Test\Semver.Test.csproj", "{EAE17B6B-5380-4017-8046-1757EB77DF53}"
-EndProject
+# Visual Studio 14
+VisualStudioVersion = 14.0.24720.0
+MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{AC069663-5FE9-438A-A9F4-C5293D695D98}"
 	ProjectSection(SolutionItems) = preProject
+		global.json = global.json
 		License.txt = License.txt
 		README.md = README.md
 	EndProjectSection
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{6CC7B594-1EBF-4FE9-8B50-8E45CB5EDAE0}"
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Semver", "src\Semver\Semver.xproj", "{63FB2390-9587-4138-995C-C634286D9071}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{5D068333-67EF-496E-AF1B-F30D1FF9AE1F}"
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Semver.Test", "tests\Semver.Test\Semver.Test.xproj", "{52307FCB-D317-4413-8FC5-A7343745A936}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,16 +24,20 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{95B2542C-E371-4AB4-9E2D-B2665CBBD629}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{95B2542C-E371-4AB4-9E2D-B2665CBBD629}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{95B2542C-E371-4AB4-9E2D-B2665CBBD629}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{95B2542C-E371-4AB4-9E2D-B2665CBBD629}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EAE17B6B-5380-4017-8046-1757EB77DF53}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EAE17B6B-5380-4017-8046-1757EB77DF53}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EAE17B6B-5380-4017-8046-1757EB77DF53}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EAE17B6B-5380-4017-8046-1757EB77DF53}.Release|Any CPU.Build.0 = Release|Any CPU
+		{63FB2390-9587-4138-995C-C634286D9071}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{63FB2390-9587-4138-995C-C634286D9071}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{63FB2390-9587-4138-995C-C634286D9071}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{63FB2390-9587-4138-995C-C634286D9071}.Release|Any CPU.Build.0 = Release|Any CPU
+		{52307FCB-D317-4413-8FC5-A7343745A936}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{52307FCB-D317-4413-8FC5-A7343745A936}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{52307FCB-D317-4413-8FC5-A7343745A936}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{52307FCB-D317-4413-8FC5-A7343745A936}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{63FB2390-9587-4138-995C-C634286D9071} = {6CC7B594-1EBF-4FE9-8B50-8E45CB5EDAE0}
+		{52307FCB-D317-4413-8FC5-A7343745A936} = {5D068333-67EF-496E-AF1B-F30D1FF9AE1F}
 	EndGlobalSection
 EndGlobal
