--- Semver/Properties/AssemblyInfo.cs
@@ -1,12 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyDescription("A semantic version implementation.")]
-[assembly: AssemblyCompany("Max Hauser")]
-[assembly: AssemblyProduct("Semver")]
-[assembly: AssemblyCopyright("Copyright © 2013 Max Hauser")]
-[assembly: ComVisible(false)]
-[assembly: Guid("e208ca67-5b59-45d9-a29a-7f30137d3beb")]
-[assembly: AssemblyVersion("1.1.2")]
-[assembly: AssemblyFileVersion("1.1.2")]
-[assembly: AssemblyInformationalVersion("1.1.2")]
