+++ tests/Semver.Test/project.json
@@ -1,0 +1,19 @@
+﻿{
+  "version": "1.0.0",
+  "description": "Semver.Test Console Application",
+  "authors": [ "Dan" ],
+  "tags": [ "" ],
+  "projectUrl": "",
+  "licenseUrl": "",
+  "dependencies": {
+    "xunit": "2.1.0",
+    "xunit.runner.dnx": "2.1.0-rc1-build204",
+    "Semver": ""
+  },
+  "commands": {
+    "test": "xunit.runner.dnx"
+  },
+  "frameworks": {
+    "dnx451": { }
+  }
+}
